refacotr: 함수 분리 및 각 뷰에 따른 Manager 생성
